moved config directory.
